requirements.txt
pyserial == 2.6
appscript == 1.0.1
